+++ packages/gcode-bgcode/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chestnutlabs/gcode-bgcode
 
+## 0.10.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chestnutlabs/gcode-containers@0.10.0
+  - @chestnutlabs/toolpath-core@0.10.0
+
 ## 0.9.0
 
 ### Patch Changes