+++ packages/gcode-colors/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chestnutlabs/gcode-colors
 
+## 0.10.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chestnutlabs/toolpath-core@0.10.0
+
 ## 0.9.0
 
 ### Patch Changes